+++ build_runtime_rk.sh
@@ -68,6 +68,9 @@ mkdir -p $OUTDIR
 cp -f mono/mini/.libs/libmono.a $OUTDIR
 cp -f mono/mini/.libs/libmono.so $OUTDIR
 
+if [ -d builds/monodistribution ] ; then
 rm -r builds/monodistribution
+fi
+
